api: fix NVS config
We were sending the NVS config in a JSON object named "config", so we
would save this to willow.json:

{
        "WAS":  {
                "URL":  "wss://was.local"
        },
        "WIFI": {
                "PSK":  "mypassword",
                "SSID": "myssid"
        }
}

That's obviously not what we want.

Rename the build_config_msg to build_msg and add a container argument to
specify in which JSON object we will put it. This way the NVS message
will currently be completely ignored by Willow until we implement
handling of the NVS message.