ui: use tabs instead of pages
The sidebar takes away a lot of screen estate. While resizing it is
possible, it can not be done directly from python code. It can be done
with st.markdown, but this requires `unsafe_allow_html=True` and I don't
like the sound of that.

Let's try tab layout instead.