ui: filter clients with unknown hostname or hw type
When connecting some WebSocket client for testing, this client shows up
in the clients overview with hostname and hw_type unknown. Hide such
clients.

If someone flashed Willow on unsupported hardware, the hostname will be
set, and the hardware type will not be unknown, so we still display it,
with hardware type "Hardware not supported.".