api: add /api/ha_url to get HA URL
We want to rip out the HA related symbols from Willow's Kconfig, but
this would break the generate_commands script.

As a first intermediary, add a way to get the HA URL from WAS, so we can
update the generate_commands to get the URL from WAS. Afterwards we can
rip out the HA Kconfig symbols.

Eventually we want to allow the user to fetch the HA entities in WAS,
save them, and allow modifying them. Once we have this functionality,
the generate script can query the Multinet command list from WAS.