utils.sh: generate timezone json
Download gen-tz.py from https://github.com/nayarsystems/posix_tz_db and
use it to generate tz.json. We don't want to include this script as it's
not using the same license as we do. By downloading it locally in
utils.sh we don't include it in our repo, nor in the Docker container,
and we don't have to worry about licensing.