ui: use selectbox for tz with values from tz.json
There doesn't seem to be a straightforward way to feed a streamlit
selectbox from a dict, using the keys as labels and the values as value.
Additionally, we cannot lookup a key based on the value, because there
are multiple timezones using an identical POSIX timezone string.

Saving the original timezone in Continent/City form is the only solution
I can come up with. Once we use a database, we can save this somewhere
else, but for now this will have to do.