api: allow post_config/post_nvs without actual config
To be able to add "Apply config" and "Apply NVS" buttons in the client
tab, we need to be able to do a POST without the actual config but with
a hostname instead. Modify post_config and post_nvs to allow this.

There's a lot of duplicate code here but FastAPI has been giving me hard
times getting this working even without further abstraction. This should
be refactored later, but there's really no time for that right now.