shared: add function to get local IP
To guess the WAS URL when we're running in Docker without starting it
via utils.sh, we instruct the user to run the container in host
networking mode. This way, we can try guessing the IP in a similar way
as we do in utils.sh. Add a function for this.