treewide: show local releases
Aside from Github releases, it should be possible to flash self-built
OTA images. Add a function to check for OTA files in the base ota
directory and add those in the releases dict. When doing an OTA from a
local build, we should skip caching it.

When there is a local OTA file for at least one hardware type, but not
all hardware types, we hit a KeyError if we use the original releases
dict to populate the dropdown. To avoid this we need a deepcopy of the
releases dict, and drop the local release if there are no binaries for
the hardware type.