api: fix /api/releases
Save the releases as a property of the FastAPI object, to fix the
following error:

UnboundLocalError: cannot access local variable 'releases' where it is not associated with a value

While at it, name it gh_releases as this is a more descriptive name, and
move it to the startup_event function. This makes it more clear this
runs at startup.