treewide: move get_releases() logic to api.py
We want to expose our internal releases dict via the API, but as we're
currently calling one of our own API endpoints in shared.get_releases(),
we need to refactor things, as we shouldn't call our own API endpoints.

Move the shared.get_releases() code to api.py, and add a new endpoint
for it. Name it get_releases_local, to describe it's our internal
release dict.

The new UI can now get our internal release dict on
/api/releases/internal/, which includes info about it being cached or
not.