README: drop main tag from container image
Since WAS NG we have a new UI. This UI is in a different repository, and
is only included in the container image on git tags. When building from
a tag, CI marks the container image with tag latest, so as of now that's
the tag people should use. Drop the tag altogether, which makes it
default to 'latest'.