app: handle all exceptions in websocket_endpoint
Connmgr sometimes ends up with the same Willow duplicated. This breaks
communication between WAS and the duplicated Willow device, and causes
nasty tracebacks. WAS has to be restarted to recover from this.

We're currently only handling ConnectionClosed and WebSocketDisconnect
exceptions in websocket_endpoint, and it appears when an unhandled
exception occurs in a WebSocket route, the client is being disconnected.
When this happens, the client is not removed from ConnMgr.

Simply handling all exceptions avoids the client from being
disconnected, and thus also avoids that client being duplicated in
ConnMgr.