app: handle all exceptions in websocket_endpoint
Connmgr sometimes ends up with the same Willow duplicated. This breaks
communication between WAS and the duplicated Willow device, and causes
nasty tracebacks. WAS has to be restarted to recover from this.

We're currently only handling ConnectionClosed and WebSocketDisconnect
exceptions in websocket_endpoint, so if we're missing some other
exception, print it so we can further debug what's going on next time we
see this problem.