treewide: support wis_tts_url_v2
Willow currently constructs the WIS TTS URL by appending
"?format=WAV&speaker=CLB&text=" to wis_tts_url in willow.json. This
makes it impossible to specify arbitrary URL parameters in WAS.

Add support for wis_tts_url_v2 to allow users to specify arbitrary URL
parameters in WAS. As the text parameter should always be at the end, we
will check if the user set it, and we will drop it and append it at the
end.

As we don't want to change the UI for this, we need to reverse the
operation in GET /api/config.