treewide: get user config from db
We exclude config entries with value None, to avoid having those written
in the JSON config on the Willow device. Because of this, we need to add
an extra check in the command endpoint code for the REST Auth Type, as
"None" is one of the valid options there.

We can simply skip setting the auth type in case, as RestAuthType.NONE
is the default for the REST command endpoint.