app: add pydantic model for user config
We're saving all the config values in the database as a string, to avoid
having to update the database schema every time we want to add a new
config option. This means we need to make sure to conver the string to
the correct value. Use a pydantic model for this.