WIP: app: add GET /api/model
This will be used by the UI to get the available wakenet models, in
preparation of dynamically updating the supported wake words on Willow.

For development I'm testing with a local JSON file and manually
generated tarballs. JSON content:

[
  {
    "friendly_name": "Alexa",
    "name": "wn9_alexa",
    "url": "http://was.dev.willow/wn9_alexa.tar.gz"
  },
  {
    "friendly_name": "Computer",
    "name": "wn9_computer_tts",
    "url": "http://was.dev.willow/wn9_computer_tts.tar.gz"
  }
]

Example content of a tarball:

drwxr-xr-x stijn/users       0 2023-05-10 05:06 wn9_alexa/
-rw-r--r-- stijn/users     168 2023-05-10 05:06 wn9_alexa/_MODEL_INFO_
-rw-r--r-- stijn/users  289638 2023-05-10 05:06 wn9_alexa/wn9_data
-rw-r--r-- stijn/users    1200 2023-05-10 05:06 wn9_alexa/wn9_index
-rw-r--r-- stijn/users       6 2023-12-15 23:23 wn9_alexa.txt

The txt file contains the friendly form of the wake word, for the
example above that would be "Alexa". We need this for the Willow help
text: "Say 'Alexa' to start!". We currently hardcode this full text in
the Willow C code, but only for 3 wake words. This will break once we
flash a wake word that is not included in the model partition during
build/flash.

We should also keep this friendly name in the JSON so that the UI can
just read that property to fill the dropdown.

TODO:
* implement getting JSON and tarball in CF worker
* add version in JSON - we can parse this from _MODEL_INFO_ but can we
  get the version of the models using an ESP-SR API?