routers.release: drop cache action from POST /api/release
CodeQL reports a possible SSRF due to not validating willow_url in the
POST data. As this API endpoint is currently not used, just remove the
cache action entirely rather than implementing validation.

Release caching happens automatically in GET /api/ota anyway, and if
anything/anyone was actually using this we can always add it back.