main: always initialize app.command_endpoint to None
When WAS command mode is disabled, init_command_endpoint doesn't always
throw an exception, so we don't always set app.command_endpoint. This
can result in the follow exception in the WebSocket loop:

was[2448155]: 2025-04-30 07:57:50 ERROR    unhandled exception in WebSocket route: 'FastAPI' object has no attribute 'command_endpoint'

This in turn can result in a connection not being cleaned up from
connmgr and possibly a duplicate client in it. If this happens, we're
seeing the following error again while trying to send a command to a
Willow:

RuntimeError: Unexpected ASGI message 'websocket.send', after sending 'websocket.close'.

Always initialize app.command_endpoint to None to fix this.