command_endpoints: remove HA REST endpoint support
When we added initial support for Home Assistant to Willow, the version
of Home Assistant we were running did not support the assist pipeline
yet, so we started with calling the REST API. When we later added
WebSocket support, we kept the REST endpoint around for people who do
not update their Home Assistant regularly.

Then when we added command endpoint support to WAS, we started with a
generic REST endpoint, and built the HA REST endpoint on top of that,
because it was easy to implement, and we could quickly validate our PoC.

It now turns out, the check for the HA assist_pipeline causes problems
when Home Assistant is unreachable during WAS startup: the check fails,
and we fall back to the HA REST endpoint. While the REST endpoint works,
it is slower compared to the WebSocket endpoint, and it does not have
the same features.

While we could do the check for assist pipeline support in a loop, this
would make the code more complicated. Keeping the HA REST endpoint
around also means we need to fix it, if something breaks in the future.
As Home Assistant supports assist pipeline for 2 years now, and to
reduce the maintenance workload, let's just remove the HA REST endpoint.